Add suport for custom configs of the topbar based on the existing
modules. These configs should be specfied in files named after the end
names of the present modules, sperated by dashes, and are stored in the
"Custom" module in the application. This module is never revealed to the
frontend, since it is removed from the module info sent to it. Two
custom configs are included to demonstrate this functionality. Bugs
might still be present.